Rename firmware/version.py -> firmware/europi_version.py, update references accordingly. Update directory structure in UF2 build docs
