+++ software/firmware/europi_version.py
